Micro refactoring (#73)
1. Remove SQLExpress DB from appsettings infrastructure (we had both Dev' and Express connection strings which is confusing)
2. Fix small typo in naming
3. Centralize and refactor semi-broken validation for the EditProfileCommand
