+++ src/UserAdmin/src/Smart.FA.Catalog.Application/Smart.FA.Catalog.Application.csproj
@@ -15,6 +15,7 @@
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.0" />
+    <PackageReference Include="Mime-Detective" Version="22.7.19" />
   </ItemGroup>
 
   <ItemGroup>
