+++ src/UserAdmin/src/Smart.FA.Catalog.Infrastructure/appsettings.Infrastructure.Development.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "Catalog": "Server=localhost\\SQLEXPRESS; Database=Catalog; Integrated Security = true"
+    "Catalog": "Server=(LocalDB)\\MSSQLLocalDB; Database=Catalog; Integrated Security=true;"
   }
 }
