+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Pages/Shared/_Layout.cshtml
@@ -1,9 +1,10 @@
-﻿@using Microsoft.AspNetCore.Localization
+﻿@using Microsoft.AspNetCore.Builder
+@using Microsoft.AspNetCore.Localization
 @using Microsoft.Extensions.Options
 @inject IOptions<RequestLocalizationOptions> _requestLocalizationOptions;
 @{
     var requestCulture = Context.Features.Get<IRequestCultureFeature>() ?? throw new InvalidOperationException($"Couldn't get {nameof(IRequestCultureFeature)}");
-    var supportedLanguges = _requestLocalizationOptions
+    var supportedLanguages = _requestLocalizationOptions
             .Value?
             .SupportedUICultures?
             .Select(supportedCulture => new SelectListItem(supportedCulture.DisplayName, supportedCulture.TwoLetterISOLanguageName)) ??
@@ -38,7 +39,7 @@
                             <select id="culture"
                                     name="culture"
                                     class="c-select"
-                                    asp-items="supportedLanguges"
+                                    asp-items="supportedLanguages"
                                     asp-for="@requestCulture.RequestCulture.UICulture.Name">
                             </select>
                         </div>
