+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Smart.FA.Catalog.Web.csproj
@@ -14,8 +14,6 @@
     <PackageReference Include="FluentValidation.AspNetCore" Version="10.3.6" />
     <PackageReference Include="Microsoft.AspNetCore.Hosting" Version="2.2.7" />
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.14.0" />
-    <PackageReference Include="Mime-Detective" Version="22.7.19" />
-    <PackageReference Include="Mime-Detective.Definitions.Condensed" Version="22.7.19" />
     <PackageReference Include="NLog" Version="4.7.14" />
     <PackageReference Include="NLog.Targets.ElasticSearch" Version="7.7.0" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="4.14.0" />
